@@ -1,6 +1,6 @@
 name: flutter_swiper
 description: The best swiper(carousel) for flutter, with multiple layouts, infinite loop. Compatible with Android & iOS.
-version: 1.1.6
+version: 2.0.0
 author: JZoom <jzoom8112@gmail.com>
 homepage: https://github.com/jzoom/flutter_swiper
 
@@ -8,15 +8,13 @@ dependencies:
   flutter:
     sdk: flutter
 
-  transformer_page_view: ^0.1.6
-  flutter_page_indicator: ^0.0.3
-
 environment:
-  sdk: ">=2.0.0-dev.48.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=0.1.4 <3.0.0"
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  pedantic: '^1.11.0'
 
 flutter:
