@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_page_indicator/flutter_page_indicator.dart';
 import 'package:flutter_swiper/flutter_swiper.dart';
 import 'config.dart';
 import 'forms/form_widget.dart';
@@ -7,48 +6,52 @@ import 'forms/form_widget.dart';
 class ExampleCustom extends StatefulWidget {
   @override
   State<StatefulWidget> createState() {
-    return new _ExampleCustomState();
+    return _ExampleCustomState();
   }
 }
 
 class _ExampleCustomState extends State<ExampleCustom> {
   //properties want to custom
-  int _itemCount;
+  late int _itemCount;
 
-  bool _loop;
+  late bool _loop;
 
-  bool _autoplay;
+  late bool _autoplay;
 
-  int _autoplayDely;
+  late int _autoplayDelay;
 
-  double _padding;
+  late double _padding;
 
-  bool _outer;
+  late bool _outer;
 
-  double _radius;
+  late double _radius;
 
-  double _viewportFraction;
+  late double _viewportFraction;
 
-  SwiperLayout _layout;
+  late SwiperLayout _layout;
 
-  int _currentIndex;
+  late int _currentIndex;
 
-  double _scale;
+  late double _scale;
 
-  Axis _scrollDirection;
+  late Axis _scrollDirection;
 
-  Curve _curve;
+  late Curve _curve;
 
-  double _fade;
+  late double _fade;
 
-  bool _autoplayDisableOnInteraction;
+  late bool _autoplayDisableOnInteraction;
 
-  CustomLayoutOption customLayoutOption;
+  late CustomLayoutOption customLayoutOption;
+
+  late SwiperController _controller;
+
+  TextEditingController numberController = TextEditingController();
 
   Widget _buildItem(BuildContext context, int index) {
     return ClipRRect(
-      borderRadius: new BorderRadius.all(new Radius.circular(_radius)),
-      child: new Image.asset(
+      borderRadius: BorderRadius.all(Radius.circular(_radius)),
+      child: Image.asset(
         images[index % images.length],
         fit: BoxFit.fill,
       ),
@@ -57,35 +60,29 @@ class _ExampleCustomState extends State<ExampleCustom> {
 
   @override
   void didUpdateWidget(ExampleCustom oldWidget) {
-    customLayoutOption = new CustomLayoutOption(startIndex: -1, stateCount: 3)
-        .addRotate([-45.0 / 180, 0.0, 45.0 / 180]).addTranslate([
-      new Offset(-370.0, -40.0),
-      new Offset(0.0, 0.0),
-      new Offset(370.0, -40.0)
-    ]);
+    customLayoutOption = CustomLayoutOption(startIndex: -1, stateCount: 3)
+        .addRotate([-45.0 / 180, 0.0, 45.0 / 180]).addTranslate(
+            [Offset(-370.0, -40.0), Offset(0.0, 0.0), Offset(370.0, -40.0)]);
     super.didUpdateWidget(oldWidget);
   }
 
   @override
   void initState() {
-    customLayoutOption = new CustomLayoutOption(startIndex: -1, stateCount: 3)
-        .addRotate([-25.0 / 180, 0.0, 25.0 / 180]).addTranslate([
-      new Offset(-350.0, 0.0),
-      new Offset(0.0, 0.0),
-      new Offset(350.0, 0.0)
-    ]);
+    customLayoutOption = CustomLayoutOption(startIndex: -1, stateCount: 3)
+        .addRotate([-25.0 / 180, 0.0, 25.0 / 180]).addTranslate(
+            [Offset(-350.0, 0.0), Offset(0.0, 0.0), Offset(350.0, 0.0)]);
     _fade = 1.0;
     _currentIndex = 0;
     _curve = Curves.ease;
     _scale = 0.8;
-    _controller = new SwiperController();
+    _controller = SwiperController();
     _layout = SwiperLayout.TINDER;
     _radius = 10.0;
     _padding = 0.0;
     _loop = true;
     _itemCount = 3;
     _autoplay = false;
-    _autoplayDely = 3000;
+    _autoplayDelay = 3000;
     _viewportFraction = 0.8;
     _outer = false;
     _scrollDirection = Axis.horizontal;
@@ -96,13 +93,13 @@ class _ExampleCustomState extends State<ExampleCustom> {
 // maintain the index
 
   Widget buildSwiper() {
-    return new Swiper(
+    return Swiper(
       onTap: (int index) {
         Navigator.of(context)
-            .push(new MaterialPageRoute(builder: (BuildContext context) {
+            .push(MaterialPageRoute(builder: (BuildContext context) {
           return Scaffold(
             appBar: AppBar(
-              title: Text("New page"),
+              title: Text('New page'),
             ),
             body: Container(),
           );
@@ -124,7 +121,7 @@ class _ExampleCustomState extends State<ExampleCustom> {
       outer: _outer,
       itemHeight: 200.0,
       viewportFraction: _viewportFraction,
-      autoplayDelay: _autoplayDely,
+      autoplayDelay: _autoplayDelay,
       loop: _loop,
       autoplay: _autoplay,
       itemBuilder: _buildItem,
@@ -132,108 +129,108 @@ class _ExampleCustomState extends State<ExampleCustom> {
       scrollDirection: _scrollDirection,
       indicatorLayout: PageIndicatorLayout.COLOR,
       autoplayDisableOnInteraction: _autoplayDisableOnInteraction,
-      pagination: new SwiperPagination(
+      pagination: SwiperPagination(
           builder: const DotSwiperPaginationBuilder(
               size: 20.0, activeSize: 20.0, space: 10.0)),
     );
   }
 
-  SwiperController _controller;
-  TextEditingController numberController = new TextEditingController();
   @override
   Widget build(BuildContext context) {
-    return new Column(children: <Widget>[
-      new Container(
+    return Column(children: <Widget>[
+      Container(
         color: Colors.black87,
-        child: new SizedBox(
+        child: SizedBox(
             height: 300.0, width: double.infinity, child: buildSwiper()),
       ),
-      new Expanded(
-          child: new ListView(
+      Expanded(
+          child: ListView(
         children: <Widget>[
-          new Text("Index:$_currentIndex"),
-          new Row(
+          Text('Index:$_currentIndex'),
+          Row(
             children: <Widget>[
-              new RaisedButton(
+              ElevatedButton(
                 onPressed: () {
                   _controller.previous(animation: true);
                 },
-                child: new Text("Prev"),
+                child: Text('Prev'),
               ),
-              new RaisedButton(
+              ElevatedButton(
                 onPressed: () {
                   _controller.next(animation: true);
                 },
-                child: new Text("Next"),
+                child: Text('Next'),
               ),
-              new Expanded(
-                  child: new TextField(
+              Expanded(
+                  child: TextField(
                 controller: numberController,
               )),
-              new RaisedButton(
+              ElevatedButton(
                 onPressed: () {
                   var text = numberController.text;
                   setState(() {
                     _currentIndex = int.parse(text);
                   });
                 },
-                child: new Text("Update"),
+                child: Text('Update'),
               ),
             ],
           ),
-          new FormWidget(
-              label: "layout",
-              child: new FormSelect(
-                  placeholder: "Select layout",
-                  value: _layout,
-                  values: [
-                    SwiperLayout.DEFAULT,
-                    SwiperLayout.STACK,
-                    SwiperLayout.TINDER,
-                    SwiperLayout.CUSTOM
-                  ],
-                  valueChanged: (value) {
-                    _layout = value;
-                    setState(() {});
-                  })),
-          new FormWidget(
-            label: "scrollDirection",
-            child: new Switch(
+          FormWidget(
+            label: 'layout',
+            child: FormSelect(
+              placeholder: 'Select layout',
+              value: _layout,
+              values: [
+                SwiperLayout.DEFAULT,
+                SwiperLayout.STACK,
+                SwiperLayout.TINDER,
+                SwiperLayout.CUSTOM
+              ],
+              valueChanged: (SwiperLayout value) {
+                _layout = value;
+                setState(() {});
+              },
+            ),
+          ),
+          FormWidget(
+            label: 'scrollDirection',
+            child: Switch(
                 value: _scrollDirection == Axis.horizontal,
                 onChanged: (bool value) => setState(() => _scrollDirection =
                     value ? Axis.horizontal : Axis.vertical)),
           ),
-          new FormWidget(
-            label: "autoplayDisableOnInteractio",
-            child: new Switch(
+          FormWidget(
+            label: 'autoplayDisableOnInteraction',
+            child: Switch(
                 value: _autoplayDisableOnInteraction,
                 onChanged: (bool value) =>
                     setState(() => _autoplayDisableOnInteraction = value)),
           ),
           //Pannel Begin
-          new FormWidget(
-            label: "loop",
-            child: new Switch(
+          FormWidget(
+            label: 'loop',
+            child: Switch(
                 value: _loop,
                 onChanged: (bool value) => setState(() => _loop = value)),
           ),
-          new FormWidget(
-            label: "outer",
-            child: new Switch(
+          FormWidget(
+            label: 'outer',
+            child: Switch(
                 value: _outer,
                 onChanged: (bool value) => setState(() => _outer = value)),
           ),
           //Pannel Begin
-          new FormWidget(
-            label: "autoplay",
-            child: new Switch(
+          FormWidget(
+            label: 'autoplay',
+            child: Switch(
                 value: _autoplay,
                 onChanged: (bool value) => setState(() => _autoplay = value)),
           ),
 
-          new FormWidget(
-            label: "padding",
-            child: new NumberPad(
+          FormWidget(
+            label: 'padding',
+            child: NumberPad(
               number: _padding,
               step: 5.0,
               min: 0.0,
@@ -244,9 +241,9 @@ class _ExampleCustomState extends State<ExampleCustom> {
               },
             ),
           ),
-          new FormWidget(
-            label: "scale",
-            child: new NumberPad(
+          FormWidget(
+            label: 'scale',
+            child: NumberPad(
               number: _scale,
               step: 0.1,
               min: 0.0,
@@ -257,9 +254,9 @@ class _ExampleCustomState extends State<ExampleCustom> {
               },
             ),
           ),
-          new FormWidget(
-            label: "fade",
-            child: new NumberPad(
+          FormWidget(
+            label: 'fade',
+            child: NumberPad(
               number: _fade,
               step: 0.1,
               min: 0.0,
@@ -270,9 +267,9 @@ class _ExampleCustomState extends State<ExampleCustom> {
               },
             ),
           ),
-          new FormWidget(
-            label: "itemCount",
-            child: new NumberPad(
+          FormWidget(
+            label: 'itemCount',
+            child: NumberPad(
               number: _itemCount,
               step: 1,
               min: 0,
@@ -284,9 +281,9 @@ class _ExampleCustomState extends State<ExampleCustom> {
             ),
           ),
 
-          new FormWidget(
-            label: "radius",
-            child: new NumberPad(
+          FormWidget(
+            label: 'radius',
+            child: NumberPad(
               number: _radius,
               step: 1.0,
               min: 0.0,
@@ -298,9 +295,9 @@ class _ExampleCustomState extends State<ExampleCustom> {
             ),
           ),
 
-          new FormWidget(
-            label: "viewportFraction",
-            child: new NumberPad(
+          FormWidget(
+            label: 'viewportFraction',
+            child: NumberPad(
               number: _viewportFraction,
               step: 0.1,
               max: 1.0,
@@ -312,23 +309,25 @@ class _ExampleCustomState extends State<ExampleCustom> {
             ),
           ),
 
-          new FormWidget(
-              label: "curve",
-              child: new FormSelect(
-                  placeholder: "Select curve",
-                  value: _layout,
-                  values: [
-                    Curves.easeInOut,
-                    Curves.ease,
-                    Curves.bounceInOut,
-                    Curves.bounceOut,
-                    Curves.bounceIn,
-                    Curves.fastOutSlowIn
-                  ],
-                  valueChanged: (value) {
-                    _curve = value;
-                    setState(() {});
-                  })),
+          FormWidget(
+            label: 'curve',
+            child: FormSelect(
+              placeholder: 'Select curve',
+              value: _layout,
+              values: [
+                Curves.easeInOut,
+                Curves.ease,
+                Curves.bounceInOut,
+                Curves.bounceOut,
+                Curves.bounceIn,
+                Curves.fastOutSlowIn
+              ],
+              valueChanged: (Curve value) {
+                _curve = value;
+                setState(() {});
+              },
+            ),
+          ),
         ],
       ))
     ]);
