@@ -1,16 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter_page_indicator/flutter_page_indicator.dart';
 import 'package:flutter_swiper/flutter_swiper.dart';
+import 'package:flutter_swiper/src/flutter_page_indicator/flutter_page_indicator.dart';
+import 'package:flutter_swiper/src/transformer_page_view/index_controller.dart';
 import 'dart:async';
 
-import 'package:transformer_page_view/transformer_page_view.dart';
+import 'package:flutter_swiper/src/transformer_page_view/transformer_page_view.dart';
 
 part 'custom_layout.dart';
 
-typedef void SwiperOnTap(int index);
+typedef SwiperOnTap = Function(int index);
 
-typedef Widget SwiperDataBuilder(BuildContext context, dynamic data, int index);
+typedef SwiperDataBuilder = Function(
+    BuildContext context, dynamic data, int index);
 
 /// default auto play delay
 const int kDefaultAutoplayDelayMs = 3000;
@@ -28,27 +30,28 @@ class Swiper extends StatefulWidget {
   final bool outer;
 
   /// Inner item height, this property is valid if layout=STACK or layout=TINDER or LAYOUT=CUSTOM,
-  final double itemHeight;
+  final double? itemHeight;
 
   /// Inner item width, this property is valid if layout=STACK or layout=TINDER or LAYOUT=CUSTOM,
-  final double itemWidth;
+  final double? itemWidth;
 
   // height of the inside container,this property is valid when outer=true,otherwise the inside container size is controlled by parent widget
-  final double containerHeight;
+  final double? containerHeight;
+
   // width of the inside container,this property is valid when outer=true,otherwise the inside container size is controlled by parent widget
-  final double containerWidth;
+  final double? containerWidth;
 
   /// Build item on index
-  final IndexedWidgetBuilder itemBuilder;
+  final IndexedWidgetBuilder? itemBuilder;
 
   /// Support transform like Android PageView did
   /// `itemBuilder` and `transformItemBuilder` must have one not null
-  final PageTransformer transformer;
+  final PageTransformer? transformer;
 
   /// count of the display items
   final int itemCount;
 
-  final ValueChanged<int> onIndexChanged;
+  final ValueChanged<int>? onIndexChanged;
 
   ///auto play config
   final bool autoplay;
@@ -74,24 +77,24 @@ class Swiper extends StatefulWidget {
   ///Index number of initial slide.
   ///If not set , the `Swiper` is 'uncontrolled', which means manage index by itself
   ///If set , the `Swiper` is 'controlled', which means the index is fully managed by parent widget.
-  final int index;
+  final int? index;
 
   ///Called when tap
-  final SwiperOnTap onTap;
+  final SwiperOnTap? onTap;
 
   ///The swiper pagination plugin
-  final SwiperPlugin pagination;
+  final SwiperPlugin? pagination;
 
   ///the swiper control button plugin
-  final SwiperPlugin control;
+  final SwiperPlugin? control;
 
   ///other plugins, you can custom your own plugin
-  final List<SwiperPlugin> plugins;
+  final List<SwiperPlugin>? plugins;
 
   ///
-  final SwiperController controller;
+  final SwiperController? controller;
 
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   ///
   final double viewportFraction;
@@ -100,53 +103,55 @@ class Swiper extends StatefulWidget {
   final SwiperLayout layout;
 
   /// this value is valid when layout == SwiperLayout.CUSTOM
-  final CustomLayoutOption customLayoutOption;
+  final CustomLayoutOption? customLayoutOption;
 
   // This value is valid when viewportFraction is set and < 1.0
-  final double scale;
+  final double? scale;
 
   // This value is valid when viewportFraction is set and < 1.0
-  final double fade;
+  final double? fade;
 
   final PageIndicatorLayout indicatorLayout;
 
   Swiper({
     this.itemBuilder,
-    this.indicatorLayout: PageIndicatorLayout.NONE,
+    this.indicatorLayout = PageIndicatorLayout.NONE,
 
     ///
     this.transformer,
-    @required this.itemCount,
-    this.autoplay: false,
-    this.layout: SwiperLayout.DEFAULT,
-    this.autoplayDelay: kDefaultAutoplayDelayMs,
-    this.autoplayDisableOnInteraction: true,
-    this.duration: kDefaultAutoplayTransactionDuration,
+    required this.itemCount,
+    this.autoplay = false,
+    this.layout = SwiperLayout.DEFAULT,
+    this.autoplayDelay = kDefaultAutoplayDelayMs,
+    this.autoplayDisableOnInteraction = true,
+    this.duration = kDefaultAutoplayTransactionDuration,
     this.onIndexChanged,
     this.index,
     this.onTap,
     this.control,
-    this.loop: true,
-    this.curve: Curves.ease,
-    this.scrollDirection: Axis.horizontal,
+    this.loop = true,
+    this.curve = Curves.ease,
+    this.scrollDirection = Axis.horizontal,
     this.pagination,
     this.plugins,
     this.physics,
-    Key key,
+    Key? key,
     this.controller,
     this.customLayoutOption,
 
     /// since v1.0.0
     this.containerHeight,
     this.containerWidth,
-    this.viewportFraction: 1.0,
+    this.viewportFraction = 1.0,
     this.itemHeight,
     this.itemWidth,
-    this.outer: false,
+    this.outer = false,
     this.scale,
     this.fade,
-  })  : assert(itemBuilder != null || transformer != null,
-            "itemBuilder and transformItemBuilder must not be both null"),
+  })  : assert(
+          itemBuilder != null || transformer != null,
+          'itemBuilder and transformItemBuilder must not be both null',
+        ),
         assert(
             !loop ||
                 ((loop &&
@@ -155,41 +160,41 @@ class Swiper extends StatefulWidget {
                             indicatorLayout == PageIndicatorLayout.COLOR ||
                             indicatorLayout == PageIndicatorLayout.NONE)) ||
                     (loop && layout != SwiperLayout.DEFAULT)),
-            "Only support `PageIndicatorLayout.SCALE` and `PageIndicatorLayout.COLOR`when layout==SwiperLayout.DEFAULT in loop mode"),
+            'Only support `PageIndicatorLayout.SCALE` and `PageIndicatorLayout.COLOR`when layout==SwiperLayout.DEFAULT in loop mode'),
         super(key: key);
 
   factory Swiper.children({
-    List<Widget> children,
-    bool autoplay: false,
-    PageTransformer transformer,
-    int autoplayDelay: kDefaultAutoplayDelayMs,
-    bool reverse: false,
-    bool autoplayDisableOnInteraction: true,
-    int duration: kDefaultAutoplayTransactionDuration,
-    ValueChanged<int> onIndexChanged,
-    int index,
-    SwiperOnTap onTap,
-    bool loop: true,
-    Curve curve: Curves.ease,
-    Axis scrollDirection: Axis.horizontal,
-    SwiperPlugin pagination,
-    SwiperPlugin control,
-    List<SwiperPlugin> plugins,
-    SwiperController controller,
-    Key key,
-    CustomLayoutOption customLayoutOption,
-    ScrollPhysics physics,
-    double containerHeight,
-    double containerWidth,
-    double viewportFraction: 1.0,
-    double itemHeight,
-    double itemWidth,
-    bool outer: false,
-    double scale: 1.0,
+    required List<Widget> children,
+    bool autoplay = false,
+    PageTransformer? transformer,
+    int autoplayDelay = kDefaultAutoplayDelayMs,
+    bool reverse = false,
+    bool autoplayDisableOnInteraction = true,
+    int duration = kDefaultAutoplayTransactionDuration,
+    ValueChanged<int>? onIndexChanged,
+    int? index,
+    SwiperOnTap? onTap,
+    bool loop = true,
+    Curve curve = Curves.ease,
+    Axis scrollDirection = Axis.horizontal,
+    SwiperPlugin? pagination,
+    SwiperPlugin? control,
+    List<SwiperPlugin>? plugins,
+    SwiperController? controller,
+    Key? key,
+    CustomLayoutOption? customLayoutOption,
+    ScrollPhysics? physics,
+    double? containerHeight,
+    double? containerWidth,
+    double viewportFraction = 1.0,
+    double? itemHeight,
+    double? itemWidth,
+    bool outer = false,
+    double scale = 1.0,
   }) {
-    assert(children != null, "children must not be null");
+    // assert(children != null, "children must not be null");
 
-    return new Swiper(
+    return Swiper(
         transformer: transformer,
         customLayoutOption: customLayoutOption,
         containerHeight: containerHeight,
@@ -222,36 +227,36 @@ class Swiper extends StatefulWidget {
   }
 
   factory Swiper.list({
-    PageTransformer transformer,
-    List list,
-    CustomLayoutOption customLayoutOption,
-    SwiperDataBuilder builder,
-    bool autoplay: false,
-    int autoplayDelay: kDefaultAutoplayDelayMs,
-    bool reverse: false,
-    bool autoplayDisableOnInteraction: true,
-    int duration: kDefaultAutoplayTransactionDuration,
-    ValueChanged<int> onIndexChanged,
-    int index,
-    SwiperOnTap onTap,
-    bool loop: true,
-    Curve curve: Curves.ease,
-    Axis scrollDirection: Axis.horizontal,
-    SwiperPlugin pagination,
-    SwiperPlugin control,
-    List<SwiperPlugin> plugins,
-    SwiperController controller,
-    Key key,
-    ScrollPhysics physics,
-    double containerHeight,
-    double containerWidth,
-    double viewportFraction: 1.0,
-    double itemHeight,
-    double itemWidth,
-    bool outer: false,
-    double scale: 1.0,
+    PageTransformer? transformer,
+    List? list,
+    CustomLayoutOption? customLayoutOption,
+    SwiperDataBuilder? builder,
+    bool autoplay = false,
+    int autoplayDelay = kDefaultAutoplayDelayMs,
+    bool reverse = false,
+    bool autoplayDisableOnInteraction = true,
+    int duration = kDefaultAutoplayTransactionDuration,
+    ValueChanged<int>? onIndexChanged,
+    int? index,
+    SwiperOnTap? onTap,
+    bool loop = true,
+    Curve curve = Curves.ease,
+    Axis scrollDirection = Axis.horizontal,
+    SwiperPlugin? pagination,
+    SwiperPlugin? control,
+    List<SwiperPlugin>? plugins,
+    SwiperController? controller,
+    Key? key,
+    ScrollPhysics? physics,
+    double? containerHeight,
+    double? containerWidth,
+    double viewportFraction = 1.0,
+    double? itemHeight,
+    double? itemWidth,
+    bool outer = false,
+    double scale = 1.0,
   }) {
-    return new Swiper(
+    return Swiper(
         transformer: transformer,
         customLayoutOption: customLayoutOption,
         containerHeight: containerHeight,
@@ -278,35 +283,33 @@ class Swiper extends StatefulWidget {
         plugins: plugins,
         physics: physics,
         itemBuilder: (BuildContext context, int index) {
-          return builder(context, list[index], index);
+          return builder!(context, list![index], index);
         },
-        itemCount: list.length);
+        itemCount: list!.length);
   }
 
   @override
   State<StatefulWidget> createState() {
-    return new _SwiperState();
+    return _SwiperState();
   }
 }
 
 abstract class _SwiperTimerMixin extends State<Swiper> {
-  Timer _timer;
+  Timer? _timer;
 
-  SwiperController _controller;
+  SwiperController? _controller;
 
   @override
   void initState() {
-    _controller = widget.controller;
-    if (_controller == null) {
-      _controller = new SwiperController();
-    }
-    _controller.addListener(_onController);
+    _controller = widget.controller ?? SwiperController();
+
+    _controller!.addListener(_onController);
     _handleAutoplay();
     super.initState();
   }
 
   void _onController() {
-    switch (_controller.event) {
+    switch (_controller!.event) {
       case SwiperController.START_AUTOPLAY:
         {
           if (_timer == null) {
@@ -328,9 +331,9 @@ abstract class _SwiperTimerMixin extends State<Swiper> {
   void didUpdateWidget(Swiper oldWidget) {
     if (_controller != oldWidget.controller) {
       if (oldWidget.controller != null) {
-        oldWidget.controller.removeListener(_onController);
+        oldWidget.controller!.removeListener(_onController);
         _controller = oldWidget.controller;
-        _controller.addListener(_onController);
+        _controller!.addListener(_onController);
       }
     }
     _handleAutoplay();
@@ -340,7 +343,7 @@ abstract class _SwiperTimerMixin extends State<Swiper> {
   @override
   void dispose() {
     if (_controller != null) {
-      _controller.removeListener(_onController);
+      _controller!.removeListener(_onController);
       //  _controller.dispose();
     }
 
@@ -349,7 +352,7 @@ abstract class _SwiperTimerMixin extends State<Swiper> {
   }
 
   bool _autoplayEnabled() {
-    return _controller.autoplay ?? widget.autoplay;
+    return _controller!.autoplay ?? widget.autoplay;
   }
 
   void _handleAutoplay() {
@@ -361,35 +364,33 @@ abstract class _SwiperTimerMixin extends State<Swiper> {
   }
 
   void _startAutoplay() {
-    assert(_timer == null, "Timer must be stopped before start!");
+    assert(_timer == null, 'Timer must be stopped before start!');
     _timer =
         Timer.periodic(Duration(milliseconds: widget.autoplayDelay), _onTimer);
   }
 
   void _onTimer(Timer timer) {
-    _controller.next(animation: true);
+    _controller!.next(animation: true);
   }
 
   void _stopAutoplay() {
     if (_timer != null) {
-      _timer.cancel();
+      _timer!.cancel();
       _timer = null;
     }
   }
 }
 
 class _SwiperState extends _SwiperTimerMixin {
-  int _activeIndex;
+  int? _activeIndex;
 
-  TransformerPageController _pageController;
+  TransformerPageController? _pageController;
 
   Widget _wrapTap(BuildContext context, int index) {
-    return new GestureDetector(
+    return GestureDetector(
       behavior: HitTestBehavior.opaque,
-      onTap: () {
-        this.widget.onTap(index);
-      },
-      child: widget.itemBuilder(context, index),
+      onTap: () => widget.onTap!(index),
+      child: widget.itemBuilder!(context, index),
     );
   }
 
@@ -397,19 +398,19 @@ class _SwiperState extends _SwiperTimerMixin {
   void initState() {
     _activeIndex = widget.index ?? 0;
     if (_isPageViewLayout()) {
-      _pageController = new TransformerPageController(
-          initialPage: widget.index,
+      _pageController = TransformerPageController(
+          initialPage: widget.index ?? 0,
           loop: widget.loop,
           itemCount: widget.itemCount,
           reverse:
-              widget.transformer == null ? false : widget.transformer.reverse,
+              widget.transformer == null ? false : widget.transformer!.reverse,
           viewportFraction: widget.viewportFraction);
     }
     super.initState();
   }
 
   bool _isPageViewLayout() {
-    return widget.layout == null || widget.layout == SwiperLayout.DEFAULT;
+    return widget.layout == SwiperLayout.DEFAULT;
   }
 
   @override
@@ -418,7 +419,7 @@ class _SwiperState extends _SwiperTimerMixin {
   }
 
   bool _getReverse(Swiper widget) =>
-      widget.transformer == null ? false : widget.transformer.reverse;
+      widget.transformer == null ? false : widget.transformer!.reverse;
 
   @override
   void didUpdateWidget(Swiper oldWidget) {
@@ -430,8 +431,8 @@ class _SwiperState extends _SwiperTimerMixin {
               widget.itemCount != oldWidget.itemCount ||
               widget.viewportFraction != oldWidget.viewportFraction ||
               _getReverse(widget) != _getReverse(oldWidget))) {
-        _pageController = new TransformerPageController(
-            initialPage: widget.index,
+        _pageController = TransformerPageController(
+            initialPage: widget.index ?? 0,
             loop: widget.loop,
             itemCount: widget.itemCount,
             reverse: _getReverse(widget),
@@ -441,7 +442,7 @@ class _SwiperState extends _SwiperTimerMixin {
       scheduleMicrotask(() {
         // So that we have a chance to do `removeListener` in child widgets.
         if (_pageController != null) {
-          _pageController.dispose();
+          _pageController!.dispose();
           _pageController = null;
         }
       });
@@ -451,17 +452,17 @@ class _SwiperState extends _SwiperTimerMixin {
     }
   }
 
-  void _onIndexChanged(int index) {
+  void _onIndexChanged(int? index) {
     setState(() {
       _activeIndex = index;
     });
     if (widget.onIndexChanged != null) {
-      widget.onIndexChanged(index);
+      widget.onIndexChanged!(index!);
     }
   }
 
   Widget _buildSwiper() {
-    IndexedWidgetBuilder itemBuilder;
+    IndexedWidgetBuilder? itemBuilder;
     if (widget.onTap != null) {
       itemBuilder = _wrapTap;
     } else {
@@ -469,7 +470,7 @@ class _SwiperState extends _SwiperTimerMixin {
     }
 
     if (widget.layout == SwiperLayout.STACK) {
-      return new _StackSwiper(
+      return _StackSwiper(
         loop: widget.loop,
         itemWidth: widget.itemWidth,
         itemHeight: widget.itemHeight,
@@ -483,13 +484,13 @@ class _SwiperState extends _SwiperTimerMixin {
         scrollDirection: widget.scrollDirection,
       );
     } else if (_isPageViewLayout()) {
-      PageTransformer transformer = widget.transformer;
+      var transformer = widget.transformer;
       if (widget.scale != null || widget.fade != null) {
         transformer =
-            new ScaleAndFadeTransformer(scale: widget.scale, fade: widget.fade);
+            ScaleAndFadeTransformer(scale: widget.scale, fade: widget.fade);
       }
 
-      Widget child = new TransformerPageView(
+      Widget child = TransformerPageView(
         pageController: _pageController,
         loop: widget.loop,
         itemCount: widget.itemCount,
@@ -497,7 +498,7 @@ class _SwiperState extends _SwiperTimerMixin {
         transformer: transformer,
         viewportFraction: widget.viewportFraction,
         index: _activeIndex,
-        duration: new Duration(milliseconds: widget.duration),
+        duration: Duration(milliseconds: widget.duration),
         scrollDirection: widget.scrollDirection,
         onPageChanged: _onIndexChanged,
         curve: widget.curve,
@@ -505,8 +506,7 @@ class _SwiperState extends _SwiperTimerMixin {
         controller: _controller,
       );
       if (widget.autoplayDisableOnInteraction && widget.autoplay) {
-        return new NotificationListener(
-          child: child,
+        return NotificationListener(
           onNotification: (ScrollNotification notification) {
             if (notification is ScrollStartNotification) {
               if (notification.dragDetails != null) {
@@ -519,12 +519,13 @@ class _SwiperState extends _SwiperTimerMixin {
 
             return false;
           },
+          child: child,
         );
       }
 
       return child;
     } else if (widget.layout == SwiperLayout.TINDER) {
-      return new _TinderSwiper(
+      return _TinderSwiper(
         loop: widget.loop,
         itemWidth: widget.itemWidth,
         itemHeight: widget.itemHeight,
@@ -538,9 +539,9 @@ class _SwiperState extends _SwiperTimerMixin {
         scrollDirection: widget.scrollDirection,
       );
     } else if (widget.layout == SwiperLayout.CUSTOM) {
-      return new _CustomLayoutSwiper(
+      return _CustomLayoutSwiper(
         loop: widget.loop,
-        option: widget.customLayoutOption,
+        option: widget.customLayoutOption!,
         itemWidth: widget.itemWidth,
         itemHeight: widget.itemHeight,
         itemCount: widget.itemCount,
@@ -553,28 +554,27 @@ class _SwiperState extends _SwiperTimerMixin {
         scrollDirection: widget.scrollDirection,
       );
     } else {
-      return new Container();
+      return SizedBox.shrink();
     }
   }
 
-  SwiperPluginConfig _ensureConfig(SwiperPluginConfig config) {
-    if (config == null) {
-      config = new SwiperPluginConfig(
-          outer: widget.outer,
-          itemCount: widget.itemCount,
-          layout: widget.layout,
-          indicatorLayout: widget.indicatorLayout,
-          pageController: _pageController,
-          activeIndex: _activeIndex,
-          scrollDirection: widget.scrollDirection,
-          controller: _controller,
-          loop: widget.loop);
-    }
+  SwiperPluginConfig _ensureConfig(SwiperPluginConfig? config) {
+    config ??= SwiperPluginConfig(
+        outer: widget.outer,
+        itemCount: widget.itemCount,
+        layout: widget.layout,
+        indicatorLayout: widget.indicatorLayout,
+        pageController: _pageController,
+        activeIndex: _activeIndex,
+        scrollDirection: widget.scrollDirection,
+        controller: _controller!,
+        loop: widget.loop);
+
     return config;
   }
 
-  List<Widget> _ensureListForStack(
-      Widget swiper, List<Widget> listForStack, Widget widget) {
+  List<Widget>? _ensureListForStack(
+      Widget swiper, List<Widget>? listForStack, Widget widget) {
     if (listForStack == null) {
       listForStack = [swiper, widget];
     } else {
@@ -585,19 +585,19 @@ class _SwiperState extends _SwiperTimerMixin {
 
   @override
   Widget build(BuildContext context) {
-    Widget swiper = _buildSwiper();
-    List<Widget> listForStack;
-    SwiperPluginConfig config;
+    var swiper = _buildSwiper();
+    List<Widget>? listForStack;
+    SwiperPluginConfig? config;
     if (widget.control != null) {
       //Stack
       config = _ensureConfig(config);
       listForStack = _ensureListForStack(
-          swiper, listForStack, widget.control.build(context, config));
+          swiper, listForStack, widget.control!.build(context, config));
     }
 
     if (widget.plugins != null) {
       config = _ensureConfig(config);
-      for (SwiperPlugin plugin in widget.plugins) {
+      for (final plugin in widget.plugins!) {
         listForStack = _ensureListForStack(
             swiper, listForStack, plugin.build(context, config));
       }
@@ -606,17 +606,17 @@ class _SwiperState extends _SwiperTimerMixin {
       config = _ensureConfig(config);
       if (widget.outer) {
         return _buildOuterPagination(
-            widget.pagination,
-            listForStack == null ? swiper : new Stack(children: listForStack),
+            widget.pagination as SwiperPagination,
+            listForStack == null ? swiper : Stack(children: listForStack),
             config);
       } else {
         listForStack = _ensureListForStack(
-            swiper, listForStack, widget.pagination.build(context, config));
+            swiper, listForStack, widget.pagination!.build(context, config));
       }
     }
 
     if (listForStack != null) {
-      return new Stack(
+      return Stack(
         children: listForStack,
       );
     }
@@ -626,42 +626,42 @@ class _SwiperState extends _SwiperTimerMixin {
 
   Widget _buildOuterPagination(
       SwiperPagination pagination, Widget swiper, SwiperPluginConfig config) {
-    List<Widget> list = [];
+    var list = <Widget>[];
     //Only support bottom yet!
     if (widget.containerHeight != null || widget.containerWidth != null) {
       list.add(swiper);
     } else {
-      list.add(new Expanded(child: swiper));
+      list.add(Expanded(child: swiper));
     }
 
-    list.add(new Align(
+    list.add(Align(
       alignment: Alignment.center,
       child: pagination.build(context, config),
     ));
 
-    return new Column(
-      children: list,
+    return Column(
       crossAxisAlignment: CrossAxisAlignment.stretch,
       mainAxisSize: MainAxisSize.min,
+      children: list,
     );
   }
 }
 
 abstract class _SubSwiper extends StatefulWidget {
-  final IndexedWidgetBuilder itemBuilder;
-  final int itemCount;
-  final int index;
-  final ValueChanged<int> onIndexChanged;
-  final SwiperController controller;
-  final int duration;
-  final Curve curve;
-  final double itemWidth;
-  final double itemHeight;
-  final bool loop;
-  final Axis scrollDirection;
+  final IndexedWidgetBuilder? itemBuilder;
+  final int? itemCount;
+  final int? index;
+  final ValueChanged<int>? onIndexChanged;
+  final SwiperController? controller;
+  final int? duration;
+  final Curve? curve;
+  final double? itemWidth;
+  final double? itemHeight;
+  final bool? loop;
+  final Axis? scrollDirection;
 
   _SubSwiper(
-      {Key key,
+      {Key? key,
       this.loop,
       this.itemHeight,
       this.itemWidth,
@@ -671,7 +671,7 @@ abstract class _SubSwiper extends StatefulWidget {
       this.controller,
       this.index,
       this.itemCount,
-      this.scrollDirection: Axis.horizontal,
+      this.scrollDirection = Axis.horizontal,
       this.onIndexChanged})
       : super(key: key);
 
@@ -680,9 +680,9 @@ abstract class _SubSwiper extends StatefulWidget {
 
   int getCorrectIndex(int indexNeedsFix) {
     if (itemCount == 0) return 0;
-    int value = indexNeedsFix % itemCount;
+    var value = indexNeedsFix % itemCount!;
     if (value < 0) {
-      value += itemCount;
+      value += itemCount!;
     }
     return value;
   }
@@ -690,18 +690,18 @@ abstract class _SubSwiper extends StatefulWidget {
 
 class _TinderSwiper extends _SubSwiper {
   _TinderSwiper({
-    Key key,
-    Curve curve,
-    int duration,
-    SwiperController controller,
-    ValueChanged<int> onIndexChanged,
-    double itemHeight,
-    double itemWidth,
-    IndexedWidgetBuilder itemBuilder,
-    int index,
-    bool loop,
-    int itemCount,
-    Axis scrollDirection,
+    Key? key,
+    Curve? curve,
+    int? duration,
+    SwiperController? controller,
+    ValueChanged<int>? onIndexChanged,
+    double? itemHeight,
+    double? itemWidth,
+    IndexedWidgetBuilder? itemBuilder,
+    int? index,
+    bool? loop,
+    int? itemCount,
+    Axis? scrollDirection,
   })  : assert(itemWidth != null && itemHeight != null),
         super(
             loop: loop,
@@ -719,24 +719,24 @@ class _TinderSwiper extends _SubSwiper {
 
   @override
   State<StatefulWidget> createState() {
-    return new _TinderState();
+    return _TinderState();
   }
 }
 
 class _StackSwiper extends _SubSwiper {
   _StackSwiper({
-    Key key,
-    Curve curve,
-    int duration,
-    SwiperController controller,
-    ValueChanged<int> onIndexChanged,
-    double itemHeight,
-    double itemWidth,
-    IndexedWidgetBuilder itemBuilder,
-    int index,
-    bool loop,
-    int itemCount,
-    Axis scrollDirection,
+    Key? key,
+    Curve? curve,
+    int? duration,
+    SwiperController? controller,
+    ValueChanged<int>? onIndexChanged,
+    double? itemHeight,
+    double? itemWidth,
+    IndexedWidgetBuilder? itemBuilder,
+    int? index,
+    bool? loop,
+    int? itemCount,
+    Axis? scrollDirection,
   }) : super(
             loop: loop,
             key: key,
@@ -753,19 +753,19 @@ class _StackSwiper extends _SubSwiper {
 
   @override
   State<StatefulWidget> createState() {
-    return new _StackViewState();
+    return _StackViewState();
   }
 }
 
 class _TinderState extends _CustomLayoutStateBase<_TinderSwiper> {
-  List<double> scales;
-  List<double> offsetsX;
-  List<double> offsetsY;
-  List<double> opacity;
-  List<double> rotates;
+  late List<double> scales;
+  late List<double?> offsetsX;
+  late List<double?> offsetsY;
+  late List<double> opacity;
+  late List<double> rotates;
 
   double getOffsetY(double scale) {
-    return widget.itemHeight - widget.itemHeight * scale;
+    return widget.itemHeight! - widget.itemHeight! * scale;
   }
 
   @override
@@ -818,30 +818,30 @@ class _TinderState extends _CustomLayoutStateBase<_TinderSwiper> {
 
   @override
   Widget _buildItem(int i, int realIndex, double animationValue) {
-    double s = _getValue(scales, animationValue, i);
-    double f = _getValue(offsetsX, animationValue, i);
-    double fy = _getValue(offsetsY, animationValue, i);
-    double o = _getValue(opacity, animationValue, i);
-    double a = _getValue(rotates, animationValue, i);
+    final s = _getValue(scales, animationValue, i)!;
+    final f = _getValue(offsetsX, animationValue, i)!;
+    final fy = _getValue(offsetsY, animationValue, i)!;
+    final o = _getValue(opacity, animationValue, i)!;
+    final a = _getValue(rotates, animationValue, i)!;
 
-    Alignment alignment = widget.scrollDirection == Axis.horizontal
+    final alignment = widget.scrollDirection == Axis.horizontal
         ? Alignment.bottomCenter
         : Alignment.centerLeft;
 
-    return new Opacity(
+    return Opacity(
       opacity: o,
-      child: new Transform.rotate(
+      child: Transform.rotate(
         angle: a / 180.0,
-        child: new Transform.translate(
-          key: new ValueKey<int>(_currentIndex + i),
-          offset: new Offset(f, fy),
-          child: new Transform.scale(
+        child: Transform.translate(
+          key: ValueKey<int>(_currentIndex + i),
+          offset: Offset(f, fy),
+          child: Transform.scale(
             scale: s,
             alignment: alignment,
-            child: new SizedBox(
+            child: SizedBox(
               width: widget.itemWidth ?? double.infinity,
               height: widget.itemHeight ?? double.infinity,
-              child: widget.itemBuilder(context, realIndex),
+              child: widget.itemBuilder!(context, realIndex),
             ),
           ),
         ),
@@ -851,9 +851,10 @@ class _TinderState extends _CustomLayoutStateBase<_TinderSwiper> {
 }
 
 class _StackViewState extends _CustomLayoutStateBase<_StackSwiper> {
-  List<double> scales;
-  List<double> offsets;
-  List<double> opacity;
+  late List<double> scales;
+  late List<double?> offsets;
+  late List<double> opacity;
+
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
@@ -861,10 +862,10 @@ class _StackViewState extends _CustomLayoutStateBase<_StackSwiper> {
 
   void _updateValues() {
     if (widget.scrollDirection == Axis.horizontal) {
-      double space = (_swiperWidth - widget.itemWidth) / 2;
+      final space = (_swiperWidth! - widget.itemWidth!) / 2;
       offsets = [-space, -space / 3 * 2, -space / 3, 0.0, _swiperWidth];
     } else {
-      double space = (_swiperHeight - widget.itemHeight) / 2;
+      final space = (_swiperHeight! - widget.itemHeight!) / 2;
       offsets = [-space, -space / 3 * 2, -space / 3, 0.0, _swiperHeight];
     }
   }
@@ -892,30 +893,30 @@ class _StackViewState extends _CustomLayoutStateBase<_StackSwiper> {
 
   @override
   Widget _buildItem(int i, int realIndex, double animationValue) {
-    double s = _getValue(scales, animationValue, i);
-    double f = _getValue(offsets, animationValue, i);
-    double o = _getValue(opacity, animationValue, i);
+    final s = _getValue(scales, animationValue, i)!;
+    final f = _getValue(offsets, animationValue, i);
+    final o = _getValue(opacity, animationValue, i)!;
 
-    Offset offset = widget.scrollDirection == Axis.horizontal
-        ? new Offset(f, 0.0)
-        : new Offset(0.0, f);
+    final offset = widget.scrollDirection == Axis.horizontal
+        ? Offset(f!, 0.0)
+        : Offset(0.0, f!);
 
-    Alignment alignment = widget.scrollDirection == Axis.horizontal
+    final alignment = widget.scrollDirection == Axis.horizontal
         ? Alignment.centerLeft
         : Alignment.topCenter;
 
-    return new Opacity(
+    return Opacity(
       opacity: o,
-      child: new Transform.translate(
-        key: new ValueKey<int>(_currentIndex + i),
+      child: Transform.translate(
+        key: ValueKey<int>(_currentIndex + i),
         offset: offset,
-        child: new Transform.scale(
+        child: Transform.scale(
           scale: s,
           alignment: alignment,
-          child: new SizedBox(
+          child: SizedBox(
             width: widget.itemWidth ?? double.infinity,
             height: widget.itemHeight ?? double.infinity,
-            child: widget.itemBuilder(context, realIndex),
+            child: widget.itemBuilder!(context, realIndex),
           ),
         ),
       ),
@@ -924,31 +925,31 @@ class _StackViewState extends _CustomLayoutStateBase<_StackSwiper> {
 }
 
 class ScaleAndFadeTransformer extends PageTransformer {
-  final double _scale;
-  final double _fade;
+  final double? _scale;
+  final double? _fade;
 
-  ScaleAndFadeTransformer({double fade: 0.3, double scale: 0.8})
+  ScaleAndFadeTransformer({double? fade = 0.3, double? scale = 0.8})
       : _fade = fade,
         _scale = scale;
 
   @override
   Widget transform(Widget item, TransformInfo info) {
-    double position = info.position;
-    Widget child = item;
+    final position = info.position;
+    var child = item;
     if (_scale != null) {
-      double scaleFactor = (1 - position.abs()) * (1 - _scale);
-      double scale = _scale + scaleFactor;
+      final scaleFactor = (1 - position!.abs()) * (1 - _scale!);
+      final scale = _scale! + scaleFactor;
 
-      child = new Transform.scale(
+      child = Transform.scale(
         scale: scale,
         child: item,
       );
     }
 
     if (_fade != null) {
-      double fadeFactor = (1 - position.abs()) * (1 - _fade);
-      double opacity = _fade + fadeFactor;
-      child = new Opacity(
+      final fadeFactor = (1 - position!.abs()) * (1 - _fade!);
+      final opacity = _fade! + fadeFactor;
+      child = Opacity(
         opacity: opacity,
         child: child,
       );
