@@ -1,5 +1,5 @@
 import 'package:flutter_swiper/src/swiper_plugin.dart';
-import 'package:transformer_page_view/transformer_page_view.dart';
+import 'package:flutter_swiper/src/transformer_page_view/index_controller.dart';
 
 class SwiperController extends IndexController {
   // Autoplay is started
@@ -16,27 +16,31 @@ class SwiperController extends IndexController {
   static const int BUILD = 5;
 
   // available when `event` == SwiperController.BUILD
-  SwiperPluginConfig config;
+  SwiperPluginConfig? config;
 
   // available when `event` == SwiperController.SWIPE
   // this value is PageViewController.pos
-  double pos;
+  double? pos;
 
-  int index;
-  bool animation;
-  bool autoplay;
+  @override
+  int? index;
+
+  @override
+  bool? animation;
+
+  bool? autoplay;
 
   SwiperController();
 
   void startAutoplay() {
     event = SwiperController.START_AUTOPLAY;
-    this.autoplay = true;
+    autoplay = true;
     notifyListeners();
   }
 
   void stopAutoplay() {
     event = SwiperController.STOP_AUTOPLAY;
-    this.autoplay = false;
+    autoplay = false;
     notifyListeners();
   }
 }
