@@ -1,15 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_swiper/flutter_swiper.dart';
-
-import 'package:flutter_page_indicator/flutter_page_indicator.dart';
+import 'package:flutter_swiper/src/flutter_page_indicator/flutter_page_indicator.dart';
 
 class FractionPaginationBuilder extends SwiperPlugin {
   ///color ,if set null , will be Theme.of(context).scaffoldBackgroundColor
-  final Color color;
+  final Color? color;
 
   ///color when active,if set null , will be Theme.of(context).primaryColor
-  final Color activeColor;
+  final Color? activeColor;
 
   ////font size
   final double fontSize;
@@ -17,51 +16,52 @@ class FractionPaginationBuilder extends SwiperPlugin {
   ///font size when active
   final double activeFontSize;
 
-  final Key key;
+  final Key? key;
 
-  const FractionPaginationBuilder(
-      {this.color,
-      this.fontSize: 20.0,
-      this.key,
-      this.activeColor,
-      this.activeFontSize: 35.0});
+  const FractionPaginationBuilder({
+    this.color,
+    this.fontSize = 20.0,
+    this.key,
+    this.activeColor,
+    this.activeFontSize = 35.0,
+  });
 
   @override
-  Widget build(BuildContext context, SwiperPluginConfig config) {
-    ThemeData themeData = Theme.of(context);
-    Color activeColor = this.activeColor ?? themeData.primaryColor;
-    Color color = this.color ?? themeData.scaffoldBackgroundColor;
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
+    final themeData = Theme.of(context);
+    final activeColor = this.activeColor ?? themeData.primaryColor;
+    final color = this.color ?? themeData.scaffoldBackgroundColor;
 
-    if (Axis.vertical == config.scrollDirection) {
-      return new Column(
+    if (Axis.vertical == config!.scrollDirection) {
+      return Column(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
-          new Text(
-            "${config.activeIndex + 1}",
+          Text(
+            '${config.activeIndex! + 1}',
             style: TextStyle(color: activeColor, fontSize: activeFontSize),
           ),
-          new Text(
-            "/",
+          Text(
+            '/',
             style: TextStyle(color: color, fontSize: fontSize),
           ),
-          new Text(
-            "${config.itemCount}",
+          Text(
+            '${config.itemCount}',
             style: TextStyle(color: color, fontSize: fontSize),
           )
         ],
       );
     } else {
-      return new Row(
+      return Row(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
-          new Text(
-            "${config.activeIndex + 1}",
+          Text(
+            '${config.activeIndex! + 1}',
             style: TextStyle(color: activeColor, fontSize: activeFontSize),
           ),
-          new Text(
-            " / ${config.itemCount}",
+          Text(
+            ' / ${config.itemCount}',
             style: TextStyle(color: color, fontSize: fontSize),
           )
         ],
@@ -72,10 +72,10 @@ class FractionPaginationBuilder extends SwiperPlugin {
 
 class RectSwiperPaginationBuilder extends SwiperPlugin {
   ///color when current index,if set null , will be Theme.of(context).primaryColor
-  final Color activeColor;
+  final Color? activeColor;
 
   ///,if set null , will be Theme.of(context).scaffoldBackgroundColor
-  final Color color;
+  final Color? color;
 
   ///Size of the rect when activate
   final Size activeSize;
@@ -86,54 +86,57 @@ class RectSwiperPaginationBuilder extends SwiperPlugin {
   /// Space between rects
   final double space;
 
-  final Key key;
+  final Key? key;
 
-  const RectSwiperPaginationBuilder(
-      {this.activeColor,
-      this.color,
-      this.key,
-      this.size: const Size(10.0, 2.0),
-      this.activeSize: const Size(10.0, 2.0),
-      this.space: 3.0});
+  const RectSwiperPaginationBuilder({
+    this.activeColor,
+    this.color,
+    this.key,
+    this.size = const Size(10.0, 2.0),
+    this.activeSize = const Size(10.0, 2.0),
+    this.space = 3.0,
+  });
 
   @override
-  Widget build(BuildContext context, SwiperPluginConfig config) {
-    ThemeData themeData = Theme.of(context);
-    Color activeColor = this.activeColor ?? themeData.primaryColor;
-    Color color = this.color ?? themeData.scaffoldBackgroundColor;
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
+    final themeData = Theme.of(context);
+    final activeColor = this.activeColor ?? themeData.primaryColor;
+    final color = this.color ?? themeData.scaffoldBackgroundColor;
 
-    List<Widget> list = [];
+    var list = <Widget>[];
 
-    if (config.itemCount > 20) {
+    if (config!.itemCount! > 20) {
       print(
-          "The itemCount is too big, we suggest use FractionPaginationBuilder instead of DotSwiperPaginationBuilder in this sitituation");
+        'The itemCount is too big, we suggest use FractionPaginationBuilder '
+        'instead of DotSwiperPaginationBuilder in this situation',
+      );
     }
 
-    int itemCount = config.itemCount;
-    int activeIndex = config.activeIndex;
+    var itemCount = config.itemCount!;
+    var activeIndex = config.activeIndex;
 
-    for (int i = 0; i < itemCount; ++i) {
-      bool active = i == activeIndex;
-      Size size = active ? this.activeSize : this.size;
+    for (var i = 0; i < itemCount; ++i) {
+      var active = i == activeIndex;
+      var size = active ? activeSize : this.size;
       list.add(SizedBox(
         width: size.width,
         height: size.height,
         child: Container(
           color: active ? activeColor : color,
-          key: Key("pagination_$i"),
+          key: Key('pagination_$i'),
           margin: EdgeInsets.all(space),
         ),
       ));
     }
 
     if (config.scrollDirection == Axis.vertical) {
-      return new Column(
+      return Column(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: list,
       );
     } else {
-      return new Row(
+      return Row(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: list,
@@ -144,10 +147,10 @@ class RectSwiperPaginationBuilder extends SwiperPlugin {
 
 class DotSwiperPaginationBuilder extends SwiperPlugin {
   ///color when current index,if set null , will be Theme.of(context).primaryColor
-  final Color activeColor;
+  final Color? activeColor;
 
   ///,if set null , will be Theme.of(context).scaffoldBackgroundColor
-  final Color color;
+  final Color? color;
 
   ///Size of the dot when activate
   final double activeSize;
@@ -158,36 +161,39 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
   /// Space between dots
   final double space;
 
-  final Key key;
+  final Key? key;
 
-  const DotSwiperPaginationBuilder(
-      {this.activeColor,
-      this.color,
-      this.key,
-      this.size: 10.0,
-      this.activeSize: 10.0,
-      this.space: 3.0});
+  const DotSwiperPaginationBuilder({
+    this.activeColor,
+    this.color,
+    this.key,
+    this.size = 10.0,
+    this.activeSize = 10.0,
+    this.space = 3.0,
+  });
 
   @override
-  Widget build(BuildContext context, SwiperPluginConfig config) {
-    if (config.itemCount > 20) {
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
+    if (config!.itemCount! > 20) {
       print(
-          "The itemCount is too big, we suggest use FractionPaginationBuilder instead of DotSwiperPaginationBuilder in this sitituation");
+        'The itemCount is too big, we suggest use FractionPaginationBuilder '
+        'instead of DotSwiperPaginationBuilder in this situation',
+      );
     }
-    Color activeColor = this.activeColor;
-    Color color = this.color;
+    var activeColor = this.activeColor;
+    var color = this.color;
 
     if (activeColor == null || color == null) {
-      ThemeData themeData = Theme.of(context);
+      final themeData = Theme.of(context);
       activeColor = this.activeColor ?? themeData.primaryColor;
       color = this.color ?? themeData.scaffoldBackgroundColor;
     }
 
     if (config.indicatorLayout != PageIndicatorLayout.NONE &&
         config.layout == SwiperLayout.DEFAULT) {
-      return new PageIndicator(
-        count: config.itemCount,
-        controller: config.pageController,
+      return PageIndicator(
+        count: config.itemCount!,
+        controller: config.pageController!,
         layout: config.indicatorLayout,
         size: size,
         activeColor: activeColor,
@@ -196,15 +202,15 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
       );
     }
 
-    List<Widget> list = [];
+    var list = <Widget>[];
 
-    int itemCount = config.itemCount;
-    int activeIndex = config.activeIndex;
+    var itemCount = config.itemCount!;
+    var activeIndex = config.activeIndex;
 
-    for (int i = 0; i < itemCount; ++i) {
-      bool active = i == activeIndex;
+    for (var i = 0; i < itemCount; ++i) {
+      var active = i == activeIndex;
       list.add(Container(
-        key: Key("pagination_$i"),
+        key: Key('pagination_$i'),
         margin: EdgeInsets.all(space),
         child: ClipOval(
           child: Container(
@@ -217,13 +223,13 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
     }
 
     if (config.scrollDirection == Axis.vertical) {
-      return new Column(
+      return Column(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: list,
       );
     } else {
-      return new Row(
+      return Row(
         key: key,
         mainAxisSize: MainAxisSize.min,
         children: list,
@@ -232,58 +238,61 @@ class DotSwiperPaginationBuilder extends SwiperPlugin {
   }
 }
 
-typedef Widget SwiperPaginationBuilder(
-    BuildContext context, SwiperPluginConfig config);
+typedef SwiperPaginationBuilder = Function(
+  BuildContext context,
+  SwiperPluginConfig? config,
+);
 
 class SwiperCustomPagination extends SwiperPlugin {
   final SwiperPaginationBuilder builder;
 
-  SwiperCustomPagination({@required this.builder}) : assert(builder != null);
+  const SwiperCustomPagination({required this.builder});
 
   @override
-  Widget build(BuildContext context, SwiperPluginConfig config) {
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
     return builder(context, config);
   }
 }
 
 class SwiperPagination extends SwiperPlugin {
   /// dot style pagination
-  static const SwiperPlugin dots = const DotSwiperPaginationBuilder();
+  static const SwiperPlugin dots = DotSwiperPaginationBuilder();
 
   /// fraction style pagination
-  static const SwiperPlugin fraction = const FractionPaginationBuilder();
+  static const SwiperPlugin fraction = FractionPaginationBuilder();
 
-  static const SwiperPlugin rect = const RectSwiperPaginationBuilder();
+  static const SwiperPlugin rect = RectSwiperPaginationBuilder();
 
   /// Alignment.bottomCenter by default when scrollDirection== Axis.horizontal
   /// Alignment.centerRight by default when scrollDirection== Axis.vertical
-  final Alignment alignment;
+  final Alignment? alignment;
 
   /// Distance between pagination and the container
   final EdgeInsetsGeometry margin;
 
-  /// Build the widet
+  /// Build the widget
   final SwiperPlugin builder;
 
-  final Key key;
+  final Key? key;
 
-  const SwiperPagination(
-      {this.alignment,
-      this.key,
-      this.margin: const EdgeInsets.all(10.0),
-      this.builder: SwiperPagination.dots});
+  const SwiperPagination({
+    this.alignment = Alignment.bottomCenter,
+    this.key,
+    this.margin = const EdgeInsets.all(10.0),
+    this.builder = SwiperPagination.dots,
+  });
 
-  Widget build(BuildContext context, SwiperPluginConfig config) {
-    Alignment alignment = this.alignment ??
-        (config.scrollDirection == Axis.horizontal
-            ? Alignment.bottomCenter
-            : Alignment.centerRight);
+  @override
+  Widget build(BuildContext context, SwiperPluginConfig? config) {
+    final alignment = config!.scrollDirection == Axis.horizontal
+        ? Alignment.bottomCenter
+        : Alignment.centerRight;
     Widget child = Container(
       margin: margin,
-      child: this.builder.build(context, config),
+      child: builder.build(context, config),
     );
-    if (!config.outer) {
-      child = new Align(
+    if (!config.outer!) {
+      child = Align(
         key: key,
         alignment: alignment,
         child: child,
