@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_page_indicator/flutter_page_indicator.dart';
 import 'package:flutter_swiper/flutter_swiper.dart';
+import 'package:flutter_swiper/src/flutter_page_indicator/flutter_page_indicator.dart';
 
 /// plugin to display swiper components
 ///
@@ -11,28 +11,27 @@ abstract class SwiperPlugin {
 }
 
 class SwiperPluginConfig {
-  final int activeIndex;
-  final int itemCount;
-  final PageIndicatorLayout indicatorLayout;
   final Axis scrollDirection;
-  final bool loop;
-  final bool outer;
-  final PageController pageController;
   final SwiperController controller;
-  final SwiperLayout layout;
+  final int? activeIndex;
+  final int? itemCount;
+  final PageIndicatorLayout? indicatorLayout;
+  final bool? loop;
+  final bool? outer;
+  final PageController? pageController;
+  final SwiperLayout? layout;
 
-  const SwiperPluginConfig(
-      {this.activeIndex,
-      this.itemCount,
-      this.indicatorLayout,
-      this.outer,
-      this.scrollDirection,
-      this.controller,
-      this.pageController,
-      this.layout,
-      this.loop})
-      : assert(scrollDirection != null),
-        assert(controller != null);
+  const SwiperPluginConfig({
+    required this.scrollDirection,
+    required this.controller,
+    this.activeIndex,
+    this.itemCount,
+    this.indicatorLayout,
+    this.outer,
+    this.pageController,
+    this.layout,
+    this.loop,
+  });
 }
 
 class SwiperPluginView extends StatelessWidget {
