@@ -4,25 +4,25 @@ import 'package:flutter_swiper/flutter_swiper.dart';
 
 void main() {
   testWidgets('Control horizontal', (WidgetTester tester) async {
-    SwiperController controller = new SwiperController();
+    final controller = SwiperController();
 
-    SwiperPluginConfig config = new SwiperPluginConfig(
+    final config = SwiperPluginConfig(
         activeIndex: 0,
         controller: controller,
         itemCount: 10,
         loop: true,
         scrollDirection: Axis.horizontal);
 
-    Key key = new UniqueKey();
-    await tester.pumpWidget(new MaterialApp(
-      home: new Scaffold(body: new Builder(builder: (BuildContext context) {
-        return new SwiperControl(key: key).build(context, config);
+    Key key = UniqueKey();
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(body: Builder(builder: (BuildContext context) {
+        return SwiperControl(key: key).build(context, config);
       })),
     ));
 
     expect(find.byKey(key), findsOneWidget);
 
-    bool first = true;
+    var first = true;
 
     await tester.tap(find.byWidgetPredicate((Widget widget) {
       if (widget is GestureDetector && first) {
@@ -35,19 +35,19 @@ void main() {
   });
 
   testWidgets('Control vertical', (WidgetTester tester) async {
-    SwiperController controller = new SwiperController();
+    final controller = SwiperController();
 
-    SwiperPluginConfig config = new SwiperPluginConfig(
+    final config = SwiperPluginConfig(
         activeIndex: 0,
         controller: controller,
         itemCount: 10,
         loop: true,
         scrollDirection: Axis.vertical);
 
-    Key key = new UniqueKey();
-    await tester.pumpWidget(new MaterialApp(
-      home: new Scaffold(body: new Builder(builder: (BuildContext context) {
-        return new SwiperControl(
+    Key key = UniqueKey();
+    await tester.pumpWidget(MaterialApp(
+      home: Scaffold(body: Builder(builder: (BuildContext context) {
+        return SwiperControl(
                 key: key, color: Colors.white, disableColor: Colors.black87)
             .build(context, config);
       })),
@@ -55,7 +55,7 @@ void main() {
 
     expect(find.byKey(key), findsOneWidget);
 
-    bool first = true;
+    var first = true;
 
     await tester.tap(find.byWidgetPredicate((Widget widget) {
       if (widget is GestureDetector && first) {
